# Conflicts:
#	.idea/workspace.xml

+++ src/image_processing/bit_depth_conversion.py
@@ -20,5 +20,5 @@ def to_8_bit(image_array, original_bit_depth=12):
     TODO Add documentation.
     """
     if original_bit_depth == 12:
-        return np.array(image_array/16, dtype=np.uint8)
+        return np.array(image_array/16, dtype=np.uint8).astype(np.uint8)
     return None
